refactor(metadata): extract CompactMetadataCard component
